+++ CHANGELOG.md
@@ -12,7 +12,9 @@ KleidiAI follows the [Semantic Versioning](https://semver.org/) specification fo
 
 - New SME micro-kernels
   - Added an x8 matmul pack micro-kernel family with 4vsx4 blocked layout, without packing bias.
+  - Added SME RHS depthwise packing kernel for FP16
 - New SME2 micro-kernels
+  - Added SME2 depthwise indirect micro-kernel for FP16.
   - kai_matmul_i32_u8p4vsx4_u8p4vsx4_i32_i32_8vsx8vs_sme2_mopa.
   - kai_matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f32_8vsx8vs_sme2_mopa.
 - Extended the following micro-kernels to support variable block length